Add user creation on startup (#16)
* Add user creation on startup

* Map device ID to running process

* Add devices per user && refactor of the layer interfaces

* Make mapping bettween measurements and user

* Add device validation for a concrete user

* Adopt new cli API version

* Add all cli resources in an archive

* Add cli binaries to the generated .zip

* Unify the units in the sql script

* Make server config files generic so they can me mounted on deploy

* Adopt mailsender from the API

* Update binaries

* Remove direct depencies

* Add checksum generation

+++ pkg/repository/query/queries.go
@@ -1,6 +1,6 @@
 package query
 
-//PostresSQL queries
+//PostreSQL queries
 const (
 	//Sensor queries
 	GetSensorNameByID       = "Select name from sensor where id=$1"
@@ -17,19 +17,23 @@ const (
 	GetSensorByName        = "SELECT id from sensor where name=$1"
 
 	//Device queries
-	GetDeviceNameByID = "Select name from device where id=$1"
-	GetDeviceIDByName = "SELECT id FROM device where name=$1"
-	InsertDevice      = `INSERT INTO device(name,description) VALUES ($1,$2)`
-	UpdateDevice      = `UPDATE device set name=$1,description=$2 where id=$3`
-	DeleteDevice      = `DELETE from device where id=$1`
+	GetDeviceNameByID = "Select d.name from device as d where d.id=$1 and d.user_id=$2"
+	GetDeviceIDByName = "SELECT id FROM device as d where d.name=$1 and d.user_id=$2"
+	InsertDevice      = `INSERT INTO device(name,description,user_id) VALUES ($1,$2,$3)`
 
-	GetDeviceByID      = `SELECT distinct d.id, d.name, d.description from device as d where d.id=$1`
-	GetAllDevices      = `SELECT distinct d.id, d.name, d.description from device_sensor as dv join device as d on dv.device_id = d.id ORDER BY d.id`
+	UpdateUserDevice = `UPDATE device set user_id=$1`
+
+	UpdateDevice = `UPDATE device set name=$1,description=$2 where id=$3 and user_id=$4`
+	DeleteDevice = `DELETE from device where id=$1 and user_id=$2`
+
+	GetDeviceByID      = `SELECT distinct d.id, d.name, d.description from device as d where d.id=$1 and d.user_id=$2`
+	GetAllDevices      = `SELECT distinct d.id, d.name, d.description from device_sensor as dv join device as d on dv.device_id = d.id where d.user_id =$1 ORDER BY d.id`
 	GetHighestDeviceID = "SELECT max(id) + 1 from device"
 
 	//User queries
 	GetHighestUserID = "SELECT max(id) + 1 from users"
 	GetUserIDByName  = "SELECT id FROM users where user_name=$1"
+	GetUserIDByEmail = "SELECT id FROM users where email=$1"
 	GetUserByID      = "SELECT s.user_name s.pass FROM user as u where ID=$1"
 	GetUserByName    = "SELECT * FROM users where user_name=$1"
 	InsertUser       = `INSERT INTO users(user_name,pass,first_name,last_name,email) VALUES ($1,$2,$3,$4,$5)`
@@ -44,13 +48,13 @@ const (
 	|> range(start: 1999-09-04T23:30:00Z)
 	|> filter(fn: (r) => r["_time"] >= %s and r["_time"] <= %s)
 	|> filter(fn: (r) => r["_measurement"] == "sensor")
-	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s")
+	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s" and r["userID"] == "%d")
 	`
 	GetAverageValueOfMeasurementsBetweenTimeStampByDeviceIdAndSensorId = `from(bucket: "%s")
 	|> range(start: 1999-09-04T23:30:00Z)
 	|> filter(fn: (r) => r["_time"] >= %s and r["_time"] <= %s)
 	|> filter(fn: (r) => r["_measurement"] == "sensor")
-	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s")
+	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s" and r["userID"] == "%d")
 	|> mean()
 	`
 
@@ -58,7 +62,7 @@ const (
 	|> range(start: 1999-09-04T23:30:00Z)
 	|> filter(fn: (r) => r["_time"] >= %s and r["_time"] <= %s)
 	|> filter(fn: (r) => r["_measurement"] == "sensor")
-	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s")
+	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s" and r["userID"] == "%d")
 	|> keep(columns: ["_value"])
 	`
 
@@ -66,13 +70,13 @@ const (
 	|> range(start: 1999-09-04T23:30:00Z)
 	|> filter(fn: (r) => r["_time"] >= %s and r["_time"] <= %s)
 	|> filter(fn: (r) => r["_measurement"] == "sensor")
-	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s")
+	|> filter(fn: (r) => r["deviceID"] == "%s" and r["sensorID"] == "%s" and r["userID"] == "%d")
 	|> count()
 	`
 
 	GetAllMeasurementsFromStartTime = `from(bucket: "%s")
 	|> range(start: 1999-09-04T23:30:00Z)
 	|> filter(fn: (r) => r["_time"] >= %s)
-	|> filter(fn: (r) => r["_measurement"] == "sensor")
+	|> filter(fn: (r) => r["_measurement"] == "sensor" and r["userID"] == "%d")
 	`
 )
